// frontend/src/app/gen/model/User.ts
/**
 * CloudRef API
 * REST API of the CloudRef system.
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface User {
    name?: string;

    firstname?: string;

    lastname?: string;

    email?: string;

    password?: string;

    userRole?: User.UserRoleEnum;

    salt?: Array<string>;

}
export namespace User {
    export enum UserRoleEnum {
        MAINTAINER = <any> 'MAINTAINER',
        USER = <any> 'USER'
    }
}